requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
async-lru==2.0.5
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
certifi==2025.6.15
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
ciscoisesdk==2.3.1
coverage==7.9.2
cryptography==45.0.5
dict2xml==1.7.6
dill==0.4.0
distro==1.9.0
f5-icontrol-rest==1.3.13
fastjsonschema==2.21.1
frozenlist==1.7.0
genie==25.5
genie.libs.clean==25.5
genie.libs.conf==25.5
genie.libs.filetransferutils==25.5
genie.libs.health==25.5
genie.libs.ops==25.5
genie.libs.parser==25.5
genie.libs.sdk==25.5
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.73.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonpickle==4.1.1
junit-xml==1.9
lxml==6.0.0
MarkupSafe==3.0.2
multidict==6.6.3
ncclient==0.6.19
netaddr==1.3.0
packaging==25.0
paramiko==3.5.1
pathspec==0.12.1
pluggy==1.6.0
ply==3.11
prettytable==3.16.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.0
pyasynchat==1.0.4
pyasyncore==1.0.4
pyats==25.5
pyats.aereport==25.5
pyats.aetest==25.5
pyats.async==25.5
pyats.connections==25.5
pyats.datastructures==25.5
pyats.easypy==25.5
pyats.kleenex==25.5
pyats.log==25.5
pyats.reporter==25.5
pyats.results==25.5
pyats.tcl==25.5
pyats.topology==25.5
pyats.utils==25.5
pycparser==2.22
pyftpdlib==2.0.1
Pygments==2.19.2
PyNaCl==1.5.0
pysmi==1.6.2
pysnmp==6.1.4
pysnmpcrypto==0.0.4
pytest==8.4.1
pytest-cov==4.1.0
python-dotenv==1.1.1
python-engineio==4.9.1
python-socketio==5.11.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rest.connector==25.5
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
tftpy==0.8.0
tqdm==4.67.1
typing_extensions==4.14.1
unicon==25.5
unicon.plugins==25.5
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
wsproto==1.2.0
xmltodict==0.12.0
yamllint==1.37.1
yang.connector==25.5
yarl==1.20.1
